refactor: extract trust evidence from the /diff page
The anonymous /diff page carried ~460 lines of provenance rendering and
off-site link building alongside its own page shell and workbench. Both
trust claims a release makes — where the bytes resolved from, and who
built them — now live in Diff/TrustEvidence.tsx (1182 -> 730 lines).

The link builders stay exported: every URL there points off-site, and a
wrong one sends a reader somewhere that looks authoritative and is not.
Their tests now import from that module directly rather than reaching
through the lazily-loaded page entry.
