+++ test/diff-evidence-links.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from "vitest";
-import { attestationLinks, githubRepoUrl, resolutionLinks } from "../src/pages/Diff";
+import { attestationLinks, githubRepoUrl, resolutionLinks } from "../src/pages/Diff/TrustEvidence";
 
 // Every value linked from the trust card is publisher-controlled: the
 // resolution trail is unsigned by construction, and even the attestation's
